fix: clean up console errors
- Rename craft-dark/craft-light Shiki themes to kata-dark/kata-light
- Centralize theme registration in shiki-themes.ts to prevent duplicate warnings
- Add wasm-unsafe-eval to CSP so Shiki WebAssembly highlighting works
- Fix Select controlled/uncontrolled warning in ModelSelector
- Downgrade unsupported extension_ui_request log from warn to debug
